requirements.txt
pandas
numpy
nltk
scikit-learn
python-box
PyYAML
fastapi
uvicorn[standard]
streamlit
wordcloud
matplotlib